requirements.txt
certifi==2018.1.18
chardet==3.0.4
idna==2.6
python-dateutil==2.6.1
requests==2.18.4
rethinkdb==2.3.0.post6
six==1.11.0
urllib3==1.22
